pyproject.toml
[tool.poetry]
name = "bluesky-llm-posting"
version = "0.1.0"
description = "Automated Bluesky Posting Bot"
authors = ["jpotter80 <jpotter80@proton.me>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12.3,<3.13"
transformers = "^4.41.1"
requests = "^2.32.2"
schedule = "^1.2.1"
pillow = "^10.3.0"
atproto = "^0.0.47"
torch = "^2.3.0"
openai = "^1.30.2"
huggingface-hub = "^0.23.1"
datetime = "^5.5"
zope-interface = "^6.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

